# docs/swagger.json
{
    "swagger": "2.0",
    "host": "",
    "basePath": "",
...
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/user/{id}": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete User",
                "description": "Delete user Superadmin",
                "operationId": "deleteUser",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "Authorization",
                        "description": "JWT token to be passed as a Header",
                        "type": "string",
                        "required": true
                    },
                    {
                        "in": "path",
                        "name": "id",
                        "description": "User MongoDB ID",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/survey/{id}": {
            "get": {
                "tags": [
                    "Survey"
                ],
                "summary": "Get All Survey List",
                "description": "Get All Survey list to List on Admin",
                "operationId": "getSurvey",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "Authorization",
                        "description": "JWT token to be passed as a Header",
                        "type": "string",
                        "required": true
                    },
                    {
                        "in": "path",
                        "name": "id",
                        "description": "Survey Id to be passed as String",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Survey"
                ],
                "summary": "Update Existing Survey",
                "description": "Update Survey - Need to Update by admin",
                "operationId": "updateSurvey",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "Authorization",
                        "description": "JWT token to be passed as a Header",
                        "type": "string",
                        "required": true
                    },
                    {
                        "in": "path",
                        "name": "id",
                        "description": "Survey Id to be passed as String",
                        "type": "string",
                        "required": true
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Survey Data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/createSurvey"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Survey"
                ],
                "summary": "Delete Existing Survey",
                "description": "Delete Survey - Need to be done by admin",
                "operationId": "deleteSurvey",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "Authorization",
                        "description": "JWT token to be passed as a Header",
                        "type": "string",
                        "required": true
                    },
                    {
                        "in": "path",
                        "name": "id",
                        "description": "Survey Id to be passed as String",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "createSurvey": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "question": {
                                "type": "string"
                            },
                            "type": {
                                "type": "string"
                            },
                            "option": {
                                "type": "string"
                            }
                        }
                    }
                },
                "assigne": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "createdBy": {
                    "type": "string"
                }
            }
        }
    }
}